Added a type cache and type deduplication
This will make the compiler use less memory for big programs, as well as speed up type comparison by a lot. This feature can be toggled on and off via the Use_Type_Cache constant, which will prove in handy when we will have a big program to benchmark this feature.